meta-rauc-raspberrypi: wic/bundle: use fixed rootfs size
Allocate a fixed size for the root-partitions, so that a the rootfs
has a (fixed) space to grow into.  Which is necessary to avoid bundle
installation failures because of the image in the bundle exceeding the
aprtition size, after any minor addition/change.

Also: use the rootfs.ext4 directly instead of the img-p5 which was
previously picked out of the wic image after its creation. This is
done because it is smaller in size, directly available (without any
deployment hacks) and most importantly: has the same content.

Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>